Extract FrameArray specs to their own file
